feature: refactor folder structure
