+++ AIMS/src/main/java/com/itep/hust/aimsgroup/controller/admin/AdminController.java
@@ -10,10 +10,9 @@ import com.itep.hust.aimsgroup.controller.admin.validator.AccountValidator;
 import com.itep.hust.aimsgroup.model.account.Account;
 import com.itep.hust.aimsgroup.model.account.AccountStatus;
 import com.itep.hust.aimsgroup.model.account.Role;
-import com.itep.hust.aimsgroup.service.dao.AccountDao;
-import com.itep.hust.aimsgroup.service.dao.AccountRoleDao;
-import com.itep.hust.aimsgroup.service.dao.RoleDao;
-import com.itep.hust.aimsgroup.service.email.EmailService;
+import com.itep.hust.aimsgroup.persistence.dao.AccountDao;
+import com.itep.hust.aimsgroup.persistence.dao.RoleDao;
+import com.itep.hust.aimsgroup.subsystem.email.EmailService;
 
 import java.util.List;
 public class AdminController {
