+++ AIMS/src/main/java/com/itep/hust/aimsgroup/controller/login/AccountLoginController.java
@@ -3,9 +3,8 @@ package com.itep.hust.aimsgroup.controller.login;
 import com.itep.hust.aimsgroup.model.account.Account;
 import com.itep.hust.aimsgroup.model.account.AccountStatus;
 import com.itep.hust.aimsgroup.model.account.Role;
-import com.itep.hust.aimsgroup.service.dao.AccountDao;
-import com.itep.hust.aimsgroup.service.dao.Dao;
-import com.itep.hust.aimsgroup.service.dao.RoleDao;
+import com.itep.hust.aimsgroup.persistence.dao.AccountDao;
+import com.itep.hust.aimsgroup.persistence.dao.RoleDao;
 
 import java.util.List;
