+++ AIMS/src/main/java/com/itep/hust/aimsgroup/controller/manager/validator/detail_infor/InforBookValidator.java
@@ -1,7 +1,5 @@
 package com.itep.hust.aimsgroup.controller.manager.validator.detail_infor;
 
-import com.itep.hust.aimsgroup.service.dao.sqlite.SqliteMediaDao;
-
 public class InforBookValidator{
 
 }
