+++ AIMS/src/main/java/com/itep/hust/aimsgroup/persistence/dao/AccountDao.java
@@ -1,4 +1,4 @@
-package com.itep.hust.aimsgroup.service.dao;
+package com.itep.hust.aimsgroup.persistence.dao;
 
 import com.itep.hust.aimsgroup.model.account.Account;
