+++ AIMS/src/main/java/com/itep/hust/aimsgroup/persistence/dao/OrderMediaDao.java
@@ -1,4 +1,4 @@
-package com.itep.hust.aimsgroup.service.dao;
+package com.itep.hust.aimsgroup.persistence.dao;
 
 import com.itep.hust.aimsgroup.model.order.OrderMedia;
 import javafx.util.Pair;
