+++ AIMS/src/main/java/com/itep/hust/aimsgroup/persistence/dao/sqlite/SqliteMediaDao.java
@@ -1,12 +1,12 @@
-package com.itep.hust.aimsgroup.service.dao.sqlite;
+package com.itep.hust.aimsgroup.persistence.dao.sqlite;
 
 import com.itep.hust.aimsgroup.model.media.Media;
 import com.itep.hust.aimsgroup.model.media.book.Book;
 import com.itep.hust.aimsgroup.model.media.cd.CD;
 import com.itep.hust.aimsgroup.model.media.cd.Track;
 import com.itep.hust.aimsgroup.model.media.dvd.DVD;
-import com.itep.hust.aimsgroup.service.dao.Dao;
-import com.itep.hust.aimsgroup.service.database.SqliteDatabase;
+import com.itep.hust.aimsgroup.persistence.dao.Dao;
+import com.itep.hust.aimsgroup.persistence.database.SqliteDatabase;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
