+++ AIMS/src/main/java/com/itep/hust/aimsgroup/persistence/dao/sqlite/SqliteRoleDao.java
@@ -1,9 +1,8 @@
-package com.itep.hust.aimsgroup.service.dao.sqlite;
+package com.itep.hust.aimsgroup.persistence.dao.sqlite;
 
 import com.itep.hust.aimsgroup.model.account.Role;
-import com.itep.hust.aimsgroup.service.dao.Dao;
-import com.itep.hust.aimsgroup.service.dao.RoleDao;
-import com.itep.hust.aimsgroup.service.database.SqliteDatabase;
+import com.itep.hust.aimsgroup.persistence.dao.RoleDao;
+import com.itep.hust.aimsgroup.persistence.database.SqliteDatabase;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
