+++ AIMS/src/main/java/com/itep/hust/aimsgroup/persistence/database/SqliteDatabase.java
@@ -1,4 +1,4 @@
-package com.itep.hust.aimsgroup.service.database;
+package com.itep.hust.aimsgroup.persistence.database;
 
 
 import java.io.BufferedReader;
@@ -6,7 +6,6 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
