+++ AIMS/src/main/java/com/itep/hust/aimsgroup/subsystem/banking/Banking.java
@@ -1,4 +1,4 @@
-package com.itep.hust.aimsgroup.subsystem;
+package com.itep.hust.aimsgroup.subsystem.banking;
 
 public interface Banking {
     void processTransaction(double money, Runnable onSuccess, Runnable onFailure);
