+++ AIMS/src/main/java/com/itep/hust/aimsgroup/subsystem/banking/vnpay/VNPayConfig.java
@@ -1,4 +1,4 @@
-package com.itep.hust.aimsgroup.subsystem.vnpay;
+package com.itep.hust.aimsgroup.subsystem.banking.vnpay;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
