+++ AIMS/src/main/java/com/itep/hust/aimsgroup/view/admin/AdminViewHandler.java
@@ -2,14 +2,13 @@ package com.itep.hust.aimsgroup.view.admin;
 
 import com.itep.hust.aimsgroup.controller.admin.AdminController;
 import com.itep.hust.aimsgroup.model.account.Account;
-import com.itep.hust.aimsgroup.model.account.AccountStatus;
 import com.itep.hust.aimsgroup.model.account.Role;
-import com.itep.hust.aimsgroup.service.dao.AccountDao;
-import com.itep.hust.aimsgroup.service.dao.RoleDao;
-import com.itep.hust.aimsgroup.service.dao.sqlite.SqliteAccountDao;
-import com.itep.hust.aimsgroup.service.dao.sqlite.SqliteRoleDao;
-import com.itep.hust.aimsgroup.service.email.EmailService;
-import com.itep.hust.aimsgroup.service.email.javax.JavaxEmailService;
+import com.itep.hust.aimsgroup.persistence.dao.AccountDao;
+import com.itep.hust.aimsgroup.persistence.dao.RoleDao;
+import com.itep.hust.aimsgroup.persistence.dao.sqlite.SqliteAccountDao;
+import com.itep.hust.aimsgroup.persistence.dao.sqlite.SqliteRoleDao;
+import com.itep.hust.aimsgroup.subsystem.email.EmailService;
+import com.itep.hust.aimsgroup.subsystem.email.javax.JavaxEmailService;
 import com.itep.hust.aimsgroup.util.Popup;
 import com.itep.hust.aimsgroup.util.Screen;
 import com.itep.hust.aimsgroup.view.login.LoginViewHandler;
