+++ AIMS/src/main/java/com/itep/hust/aimsgroup/view/login/AccountLoginViewHandler.java
@@ -3,8 +3,8 @@ package com.itep.hust.aimsgroup.view.login;
 import com.itep.hust.aimsgroup.controller.login.AccountLoginController;
 import com.itep.hust.aimsgroup.model.account.Account;
 import com.itep.hust.aimsgroup.model.account.Role;
-import com.itep.hust.aimsgroup.service.dao.sqlite.SqliteAccountDao;
-import com.itep.hust.aimsgroup.service.dao.sqlite.SqliteRoleDao;
+import com.itep.hust.aimsgroup.persistence.dao.sqlite.SqliteAccountDao;
+import com.itep.hust.aimsgroup.persistence.dao.sqlite.SqliteRoleDao;
 import com.itep.hust.aimsgroup.util.Popup;
 import com.itep.hust.aimsgroup.util.Screen;
 import com.itep.hust.aimsgroup.view.admin.AdminViewHandler;
